Renamed PublishingRequestResolver to TicketResolver and moved persistTicket from handler to resolver. Added functionality and tests for error handling for publishing workflow.
